Improve assign lyrics
